requirements/base.txt
Django>=1.7,<1.8
boto>=2.36.0
dj-database-url>=0.3.0
django-activeurl>=0.1.8
django-admin-tools>=0.5.2
django-allauth>=0.19.0
django-analytical>=0.19.0
django-braces>=1.4.0
django-crispy-forms>=1.4.0
django-extensions>=1.5.0
django-model-utils>=2.2.0
django-ratings>=0.3.7
-e git://github.com/danawoodman/django-reviews.git#egg=django_reviews-dev
django-secure>=1.0
django-storages>=1.1.8
feedparser>=5.1.3
gevent>=1.0
greenlet>=0.4.2
gunicorn>=0.18.0
mailchimp>=2.0.7
mandrill>=1.0.53
Pillow>=2.3.0
postmonkey>=1.0b
psycopg2>=2.5.2
raven>=4.0.4
simplejson>=3.3.3
Unipath>=1.1
